separate add_existing_infrastructure
