+++ workflow/Snakefile
@@ -15,8 +15,10 @@ from scripts._helpers import PathManager, get_cutout_params, ConfigManager
 configfile: "config/default_config.yaml"
 configfile: "config/technology_config.yaml"
 configfile: "config/plot_config.yaml"
-configfile: "config/remind.yaml"
 
+# for dev of remind with mocksnakemake, activate this
+# configfile: "config/templates/remind_cpled.yml"
+# configfile: "config/remind.yaml"
 
 config_manager = ConfigManager(config)
 config = config_manager.handle_scenarios()
@@ -45,7 +47,6 @@ include: "rules/prepare_remind_coupled.smk"
 
 if config["run"].get("is_test", False):
     localrules: plot_all, build_population, dag, fetch_region_shapes
-
 
 # ====== set up snakemake providers =======
 storage:
@@ -259,7 +260,18 @@ if config["foresight"] in ["overnight"]:
         resources: mem_mb=10000
         script: "scripts/prepare_network.py"
 
-    
+    rule prepare_baseyear_capacities:
+        input:
+            tech_costs=COSTS_DATA + "/costs_{planning_horizons}.csv",
+            **{f"{tech}": f"resources/data/existing_infrastructure/{tech} capacity.csv"
+               for tech in config['existing_capacities']['techs']},
+        # TODO make dep on remind
+        output: 
+            installed_capacities = DERIVED_DATA+ '/existing_infrastructure/capacities_{planning_horizons}.csv'
+        threads: 1
+        resources: mem_mb=2000
+        script: "scripts/prepare_existing_capacities.py"
+
     rule add_existing_baseyear:
         params:
             baseyear = config["scenario"]["planning_horizons"][0],
@@ -268,8 +280,7 @@ if config["foresight"] in ["overnight"]:
             network=RESULTS_DIR+ '/prenetworks/ntwk_{planning_horizons}.nc',
             tech_costs=COSTS_DATA + "/costs_{planning_horizons}.csv",
             cop_name=DERIVED_CUTOUT + "/heating/hp_cop.h5",
-            **{f"{tech}": f"resources/data/existing_infrastructure/{tech} capacity.csv"
-               for tech in config['existing_capacities']['techs']},
+            installed_capacities = DERIVED_DATA+ '/existing_infrastructure/capacities_{planning_horizons}.csv',
         output: RESULTS_DIR+ '/prenetworks-brownfield/ntwk_{planning_horizons}.nc'
         threads: 1
         resources: mem_mb=2000
